+++ tileops/ops/moe/__init__.py
@@ -1,24 +1,46 @@
 """MoE operator package."""
 
-from .abc import (
+from .fused_moe import FusedMoe, FusedMoeFwdCbFwdOp, FusedMoeFwdOp
+from .fused_topk import FusedTopKOp
+from .permute_align import MoePermuteAlignFwdOp
+from .prepare_finalize.no_dp_ep import MoEPrepareAndFinalizeNoDPEP
+from .routed_expert import (
     FusedMoEExperts,
     FusedMoEExpertsModular,
-    FusedMoEPrepareAndFinalize,
+    FusedMoEExpertsNopadPersistent3WGFwdOp,
+    FusedMoEExpertsPaddedFwdOp,
+    MoeGroupedGemmNopadFwdOp,
+    MoePermuteNopadFwdOp,
+    MoePermutePaddedFwdOp,
+    MoeUnpermuteFwdOp,
     PrepareResult,
     WeightedReduce,
     WeightedReduceNoOp,
 )
-from .experts.nopad import FusedMoEExpertsNopadPersistent3WGFwdOp
-from .experts.padded import FusedMoEExpertsPaddedFwdOp
-from .fused_moe import FusedMoe, FusedMoeFwdCbFwdOp, FusedMoeFwdOp
-from .fused_topk import FusedTopKOp
-from .moe_grouped_gemm_nopad import MoeGroupedGemmNopadFwdOp
-from .permute_align import MoePermuteAlignFwdOp
-from .permute_nopad import MoePermuteNopadFwdOp
-from .permute_padded import MoePermutePaddedFwdOp
-from .prepare_finalize.no_dp_ep import MoEPrepareAndFinalizeNoDPEP
+from .routed_expert.abc import FusedMoEPrepareAndFinalize
 from .shared_fused_moe import SharedFusedMoE
-from .unpermute import MoeUnpermuteFwdOp
+
+__all__ = [
+    "FusedMoEExperts",
+    "FusedMoEExpertsModular",
+    "FusedMoEExpertsNopadPersistent3WGFwdOp",
+    "FusedMoEExpertsPaddedFwdOp",
+    "FusedMoEPrepareAndFinalize",
+    "FusedMoe",
+    "FusedMoeFwdCbFwdOp",
+    "FusedMoeFwdOp",
+    "FusedTopKOp",
+    "MoEPrepareAndFinalizeNoDPEP",
+    "MoeGroupedGemmNopadFwdOp",
+    "MoePermuteAlignFwdOp",
+    "MoePermuteNopadFwdOp",
+    "MoePermutePaddedFwdOp",
+    "MoeUnpermuteFwdOp",
+    "PrepareResult",
+    "SharedFusedMoE",
+    "WeightedReduce",
+    "WeightedReduceNoOp",
+]
 
 __all__ = [
     "FusedMoEExperts",
