--- tileops/ops/moe/experts/__init__.py
@@ -1,6 +1,0 @@
-"""MoE expert GEMM implementations."""
-
-from .nopad import FusedMoEExpertsNopadPersistent3WGFwdOp
-from .padded import FusedMoEExpertsPaddedFwdOp
-
-__all__ = ["FusedMoEExpertsNopadPersistent3WGFwdOp", "FusedMoEExpertsPaddedFwdOp"]
