+++ tileops/ops/moe/prepare_finalize/no_dp_ep.py
@@ -4,7 +4,11 @@ from __future__ import annotations
 
 from torch import Tensor
 
-from tileops.ops.moe.abc import FusedMoEPrepareAndFinalize, PrepareResult, WeightedReduce
+from tileops.ops.moe.routed_expert.abc import (
+    FusedMoEPrepareAndFinalize,
+    PrepareResult,
+    WeightedReduce,
+)
 
 __all__ = ["MoEPrepareAndFinalizeNoDPEP"]
