+++ tileops/ops/moe/routed_expert/fused_routed_expert.py
@@ -1,4 +1,4 @@
-"""FusedMoEExpertsNopadPersistent3WGFwdOp — tight (no-pad) layout expert GEMM with 3WG persistent kernel."""
+"""FusedMoEExperts implementations: padded and nopad+3WG variants."""
 
 from __future__ import annotations
 
@@ -8,25 +8,162 @@ from typing import Optional
 import torch
 from torch import Tensor
 
-from tileops.kernels.grouped_gemm import GroupedGemmPersistent3WGKernel
+from tileops.kernels.grouped_gemm import (
+    _DEFAULT_CONFIGS as _GEMM_DEFAULT_CONFIGS,
+)
+from tileops.kernels.grouped_gemm import (
+    GroupedGemmPersistent3WGKernel,
+)
 from tileops.kernels.grouped_gemm.grouped_gemm_persistent_3wg import (
     _DEFAULT_CONFIG as _3WG_DEFAULT_CONFIG,
 )
 from tileops.kernels.moe.moe_grouped_gemm_nopad import MoeGroupedGemmNopadKernel
 from tileops.ops.elementwise import SiluAndMulFwdOp
-from tileops.ops.moe.abc import (
+from tileops.ops.grouped_gemm import GroupedGemmOp
+
+from .abc import (
     FusedMoEExpertsModular,
     WeightedReduce,
     WeightedReduceNoOp,
     _validate_fused_moe_experts_dtypes,
 )
-from tileops.ops.moe.moe_grouped_gemm_nopad import MoeGroupedGemmNopadFwdOp
-from tileops.ops.moe.permute_nopad import MoePermuteNopadFwdOp
-from tileops.ops.moe.unpermute import MoeUnpermuteFwdOp
+from .moe_grouped_gemm_nopad import MoeGroupedGemmNopadFwdOp
+from .permute_nopad import MoePermuteNopadFwdOp
+from .permute_padded import MoePermutePaddedFwdOp
+from .unpermute import MoeUnpermuteFwdOp
 
-__all__ = ["FusedMoEExpertsNopadPersistent3WGFwdOp"]
+__all__ = [
+    "FusedMoEExpertsNopadPersistent3WGFwdOp",
+    "FusedMoEExpertsPaddedFwdOp",
+]
 
 _logger = logging.getLogger(__name__)
+
+_BLOCK_M: int = _GEMM_DEFAULT_CONFIGS[(False, True)]["block_m"]
+
+
+class FusedMoEExpertsPaddedFwdOp(FusedMoEExpertsModular):
+    """Expert GEMM using block_m-aligned padded layout (reference baseline).
+
+    Internal pipeline: MoePermutePaddedFwdOp → gate_up GEMM → SwiGLU →
+    down GEMM → MoeUnpermuteFwdOp (weighted reduction included).
+
+    expert_map is not supported; raises NotImplementedError if non-None.
+    """
+
+    def __init__(
+        self,
+        num_tokens: int,
+        num_experts: int,
+        top_k: int,
+        hidden_size: int,
+        ffn_size: int,
+        routed_scaling_factor: float = 1.0,
+        dtype: torch.dtype = torch.bfloat16,
+        expert_map: Optional[Tensor] = None,
+        kernel_map=None,
+    ):
+        self.dispatch_kernel(kernel_map)
+        if expert_map is not None:
+            raise NotImplementedError(
+                "expert_map is not supported for padded layout. "
+                "Use FusedMoEExpertsNopadPersistent3WGFwdOp for EP mode."
+            )
+        self.num_tokens = num_tokens
+        self.num_experts = num_experts
+        self.top_k = top_k
+        self.hidden_size = hidden_size
+        self.ffn_size = ffn_size
+        self.dtype = dtype
+        numel = num_tokens * top_k
+        padded_batch_sum = numel + (num_experts * (_BLOCK_M - 1))
+
+        self._permute = MoePermutePaddedFwdOp(
+            total_tokens=num_tokens, top_k=top_k, num_experts=num_experts,
+            hidden_size=hidden_size, dtype=dtype, block_m=_BLOCK_M,
+            kernel_map=kernel_map,
+        )
+        self._gemm_gate_up = GroupedGemmOp(
+            batch_sum=padded_batch_sum, batch_count=num_experts,
+            n=ffn_size * 2, k=hidden_size, dtype=dtype,
+            kernel_map=kernel_map,
+        )
+        self._silu_and_mul = SiluAndMulFwdOp(
+            M=padded_batch_sum, N=ffn_size, dtype=dtype, kernel_map=kernel_map,
+        )
+        self._gemm_down = GroupedGemmOp(
+            batch_sum=padded_batch_sum, batch_count=num_experts,
+            n=hidden_size, k=ffn_size, dtype=dtype,
+            kernel_map=kernel_map,
+        )
+        self._unpermute = MoeUnpermuteFwdOp(
+            total_tokens=num_tokens, top_k=top_k,
+            hidden_size=hidden_size, dtype=dtype, padded_batch_sum=padded_batch_sum,
+            kernel_map=kernel_map,
+        )
+        self._routed_scaling_factor = routed_scaling_factor
+
+    def workspace_shapes(
+        self, M: int, N: int, K: int, topk: int, num_experts: int,
+    ) -> tuple[tuple[int, ...], tuple[int, ...]]:
+        return ((0,), (0,))
+
+    def output_shape(self, T_prime: int, H: int) -> tuple[int, int]:
+        return (T_prime, H)
+
+    def make_weighted_reduce(self) -> WeightedReduce:
+        return WeightedReduceNoOp()
+
+    def _validate_dtypes(
+        self,
+        output: Tensor,
+        hidden_states: Tensor,
+        w_gate_up: Tensor,
+        w_down: Tensor,
+        topk_weights: Tensor,
+        topk_ids: Tensor,
+        expert_map: Tensor | None,
+        workspace1: Tensor,
+        workspace2: Tensor,
+    ) -> None:
+        _validate_fused_moe_experts_dtypes(
+            self.dtype,
+            output, hidden_states, w_gate_up, w_down,
+            topk_weights, topk_ids, expert_map, workspace1, workspace2,
+        )
+
+    @property
+    def default_kernel_map(self) -> dict:
+        # All sub-kernels are owned by the inner Ops (permute / GEMM / SiLU / unpermute).
+        return {}
+
+    def forward(
+        self,
+        output: Tensor,
+        hidden_states: Tensor,
+        w_gate_up: Tensor,
+        w_down: Tensor,
+        topk_weights: Tensor,
+        topk_ids: Tensor,
+        expert_map: Tensor | None,
+        workspace1: Tensor,
+        workspace2: Tensor,
+        num_experts: int,
+    ) -> None:
+        perm_h_pad, padded_offsets, padded_sizes, _, fwd_idx = self._permute(
+            hidden_states, topk_ids
+        )
+        gate_up_pad = self._gemm_gate_up(
+            perm_h_pad, w_gate_up, padded_sizes, padded_offsets, padded_offsets
+        )
+        act_pad = self._silu_and_mul(gate_up_pad)
+        mm2_pad = self._gemm_down(
+            act_pad, w_down, padded_sizes, padded_offsets, padded_offsets
+        )
+        result = self._unpermute(mm2_pad, fwd_idx, topk_weights)
+        if self._routed_scaling_factor != 1.0:
+            result = result * self._routed_scaling_factor
+        output.copy_(result)
 
 
 class FusedMoEExpertsNopadPersistent3WGFwdOp(FusedMoEExpertsModular):
