+++ tileops/ops/moe/routed_expert/moe_grouped_gemm_nopad.py
@@ -7,7 +7,7 @@ import torch
 from tileops.kernels.kernel_base import Kernel
 from tileops.kernels.moe.moe_grouped_gemm_nopad import MoeGroupedGemmNopadKernel
 
-from ..op_base import Op
+from ...op_base import Op
 
 __all__ = ["MoeGroupedGemmNopadFwdOp"]
